Extract MemoryPersistentCache from CachingPersistenceTest to allow reuse in other tests.
